% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{findSubexprs}
\alias{findSubexprs}
\title{findSubexprs}
\usage{
findSubexprs(expr, simplify = FALSE, tag = ".expr", verbose = FALSE, ...)
}
\arguments{
\item{expr}{An expression represented in a variety of ways. See Details.}

\item{simplify}{The environment in which simplifications are stored.}

\item{tag}{to be attached to the returned object(s)??}

\item{verbose}{If \code{TRUE}, then diagnostic output will be printed as derivatives
and simplifications are recognized.}

\item{...}{Additional arguments}
}
\description{
Try to find the sub-expressions in \code{expr} ??
}
