% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{fitted.nlsr}
\alias{fitted.nlsr}
\title{fitted.nlsr}
\usage{
\method{fitted}{nlsr}(object = NULL, data = parent.frame(), ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{data}{a data frame with the date for which fits are wanted
?? how do we guarantee it is the data used to fit the model
Or do we need a different approach?}

\item{...}{additional data needed to evaluate the modeling functions
Default FALSE}
}
\description{
prepare and display fits of \code{nlsr} computations
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
