% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{isCALL}
\alias{isCALL}
\title{isCALL
 
Test if argument is a call}
\usage{
isCALL(x, name)
}
\arguments{
\item{x}{object to be tested}

\item{name}{??need to document better -- is it a character string?}
}
\description{
isCALL
 
Test if argument is a call
}
