% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{nvec}
\alias{nvec}
\title{nvec}
\usage{
nvec(vec)
}
\arguments{
\item{vec}{a named vector of parameters}
}
\value{
none (??may want to change that)
}
\description{
Compact display of specified \code{nvec} named vector
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
