% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{pctrl}
\alias{pctrl}
\title{pctrl}
\usage{
pctrl(control)
}
\arguments{
\item{control}{a control list}
}
\value{
none (??may want to change that)
}
\description{
Compact display of specified \code{control} vector
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
