% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{pnlm0}
\alias{pnlm0}
\title{pnlm0}
\usage{
pnlm0(x)
}
\arguments{
\item{x}{a minpack.lm nlsLM() or nls.lm() result object}
}
\value{
none
}
\description{
Compact display of specified \code{minpack.lm} object \code{x}
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
