% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{pnls0}
\alias{pnls0}
\title{pnls0}
\usage{
pnls0(x)
}
\arguments{
\item{x}{an nls() result object}
}
\value{
none
}
\description{
Compact display of specified \code{nls} object \code{x}
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
