\name{pnlslm}
\alias{pnlslm}
\title{pnlslm}
\usage{
pnlslm(x)
}
\arguments{
\item{x}{an nls() result object from minpack.lm::nls.lm()}
}
\value{
none
}
\description{
Compact display of specified \code{nls.lm} object \code{x}.
This code returns the iteration count in a different variable
from that of \code{nls} objects.
}
\author{
J C Nash 2014-7-16, 2023-5-8 nashjc _at_ uottawa.ca
}
