\name{print.nlsr}
\alias{print.nlsr}
\title{print.nlsr}
\usage{
\method{print}{nlsr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{nlsr}}

\item{...}{additional data needed to evaluate the modeling functions
Default FALSE}
}
\description{
prepare and display result of \code{nlsr} computations
}
\details{
The set of possible controls to set is as follows
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
