% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{pshort}
\alias{pshort}
\title{pshort}
\usage{
pshort(x)
}
\arguments{
\item{x}{an object of class \code{nlsr}}
}
\description{
1 line result display of \code{nlsr} computations
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
