% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsr.package.R
\name{rawres.nlsr}
\alias{rawres.nlsr}
\title{rawres.nlsr}
\usage{
rawres.nlsr(object = NULL, data = parent.frame(), ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{data}{a data frame with the date for which fits are wanted
?? how do we guarantee it is the data used to fit the model
Or do we need a different approach?}

\item{...}{additional data needed to evaluate the modeling functions
Default FALSE}
}
\value{
A vector of the raw residuals
}
\description{
prepare and display raw residuals of \code{nlsr} computations
NOTE: we use model - data i.e., rhs - lhs
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
