% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resJacFns.R
\name{resgr}
\alias{resgr}
\title{resgr}
\usage{
resgr(prm, resfn, jacfn, ...)
}
\arguments{
\item{prm}{a numeric vector of parameters to the model}

\item{resfn}{a function to compute a vector of residuals}

\item{jacfn}{a function to compute the Jacobian of the sum of squares}

\item{...}{Extra information needed to compute the residuals}
}
\description{
Computes the gradient of the sum of squares function for nonlinear least
squares where \code{resfn} and \code{jacfn} supply the residuals and Jacobian
}
\details{
?? does it work with approximate Jacobian functions
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
