\name{resid.nlsr}
\alias{resid.nlsr}
\title{resid.nlsr}
\usage{
\method{resid}{nlsr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{...}{additional data needed to evaluate the modeling functions}
}
\description{
prepare and display result of \code{nlsr} computations
}
\author{
J C Nash nashjc _at_ uottawa.ca

### remove _at_export to try to overcome NAMESPACE issue
}
