\name{residuals.nlsr}
\alias{residuals.nlsr}
\title{residuals.nlsr}
\usage{
\method{residuals}{nlsr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{...}{additional data needed to evaluate the modeling functions}
}
\description{
prepare and display result of \code{nlsr} computations
}
\author{
J C Nash nashjc _at_ uottawa.ca
}
