\name{resss}
\alias{resss}
\title{resss}
\usage{
resss(prm, resfn, ...)
}
\arguments{
\item{prm}{a named numeric vector of parameters to the model}

\item{resfn}{a function to compute a vector of residuals}

\item{...}{Extra information needed to compute the residuals}
}
\description{
compute the sum of squares from \code{resfn} at parameters \code{prm}
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
