\name{summary.nlsr}
\alias{summary.nlsr}
\title{summary.nlsr}
\usage{
\method{summary}{nlsr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{...}{additional data needed to evaluate the modeling functions}
}
\description{
prepare display result of \code{nlsr} computations - NOT compact output
}
\details{
The set of possible controls to set is as follows
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
