\docType{data}
\name{sysDerivs}
\alias{sysDerivs}
\alias{sysSimplifications}
\title{Internal Environments}
\format{
\code{environment}s holding derivatives and simplification rules.
}
\usage{
sysDerivs
sysSimplifications
}
\description{
These are the default collection of derivatives and 
simplifications used internally by
\code{\link{nlsDeriv}} and \code{\link{nlsSimplify}}.
}
\keyword{internal}
