% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\docType{data}
\name{sysSimplifications}
\alias{sysSimplifications}
\title{sysSimplifications}
\format{
An object of class \code{environment} of length 13.
}
\usage{
sysSimplifications
}
\description{
creates a new environment whose parent is emptyenv
}
\keyword{datasets}
