\name{coef.nlsr}
\alias{coef.nlsr}
\title{coef.nlsr}
\usage{
\method{coef}{nlsr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{...}{additional data needed to evaluate the modeling functions
Default FALSE}
}
\description{
prepare and display result of \code{nlsr} computations
}
\details{
The set of possible controls to set is as follows
}
\author{
J C Nash 2014-7-16   nashjc _at_ uottawa.ca
}
