% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.nlsr.R
\name{coef.nlsr}
\alias{coef.nlsr}
\title{coef.nlsr: Output model coefficients for nlsr object.}
\usage{
\method{coef}{nlsr}(object, ...)
}
\arguments{
\item{object}{An object of class 'nlsr'}

\item{...}{dot-args to provide exogenous data to the problem
Any data needed for the function. We do not know of any!}
}
\description{
A routine to extract and display the coefficients for a model 
estimated by \code{nlxb} or \code{nlfb} in the \code{nlsr} structured
\code{object}.
}
\section{}{
NA
}

\seealso{
Function \code{nls()}, packages \code{\link{optim}} and \code{optimx}.
}
\author{
John C Nash <nashjc@uottawa.ca>
}
