\name{fitted.nlsr}
\alias{fitted.nlsr}
\title{fitted.nlsr}
\usage{
\method{fitted}{nlsr}(object = NULL, data = parent.frame(), ...)
}
\arguments{
\item{object}{an object of class \code{nlsr}}

\item{data}{a data frame with the data for which fits are wanted.}

\item{...}{additional data needed to evaluate the modeling functions
Default FALSE}
}
\description{
prepare and display fits of \code{nlsr} computations
}
\author{
J C Nash 2014-7-16 revised 2022-11-22  nashjc _at_ uottawa.ca
}
