% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{newDeriv}
\alias{newDeriv}
\title{newDeriv}
\usage{
newDeriv (expr, deriv, derivEnv = sysDerivs)
}
\arguments{
\item{expr}{An expression represented in a variety of ways.}

\item{deriv}{An expression giving the derivative of the function call in \code{expr}.}

\item{derivEnv}{The environment in which to evaluate (??) these}
}
\description{
Define a new derivative expression
}
