% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{newSimplification}
\alias{newSimplification}
\title{newSimplification}
\usage{
newSimplification(expr, test, simplification, do_eval = FALSE, 
                       simpEnv = sysSimplifications)
}
\arguments{
\item{expr}{An expression represented in a variety of ways.}

\item{test}{?? what is this}

\item{simplification}{?? what is this}

\item{do_eval}{??evaluate the result}

\item{simpEnv}{the environment in which the simplification is carried out}
}
\description{
Define a new simplification expression
}
