% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{nlsSimplify}
\alias{nlsSimplify}
\title{nlsSimplify}
\usage{
nlsSimplify(expr, simpEnv = sysSimplifications, verbose = FALSE)
}
\arguments{
\item{expr}{An expression represented in a variety of ways. See Details.}

\item{simpEnv}{The environment in which simplifications are stored.}

\item{verbose}{If \code{TRUE}, then diagnostic output will be printed as derivatives
and simplifications are recognized.}
}
\description{
Try to simplify an expression re: nonlinear least squares
}
