% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\docType{data}
\name{sysDerivs}
\alias{sysDerivs}
\title{sysDerivs}
\format{
An object of class \code{environment} of length 27.
}
\usage{
sysDerivs()
}
\description{
creates a new environment whose parent is emptyenv
}
\keyword{datasets}
